requirements.txt
asgiref==3.9.1
Django==5.2.6
gunicorn==23.0.0
packaging==25.0
pillow==11.3.0
python-decouple==3.8
sqlparse==0.5.3
whitenoise==6.11.0
